Allow pruning items by reference, instead of just path.
Doing it by path would be hella ineffecient and inelegant if we're
making multiple changes to a tree in memory.  Imagine doing something
SUPER UNCOMMON like "remove all the tasks tagged with @done": if we only
use paths, we'll have to traverse the tree each time we want to remove
an item, because the paths could have changed.  If we can just remove
items by reference, then we have to traverse the tree once.