Mutate by replacement; handle mutation edge cases around roots.
This will help in "processing" (in the GTD sense) a list: one thing you
might want to do is edit an item.  The easy way to do that is: write the
item to a file, open that in your editor, then take the result and
splice it back into the thing you were processing.

In this flow, the thing you're splicing in will be a root and could have
multiple children.